Split migrate-converged-pkg into separate files
