+++ tools/generators/migrate-converged-pkg/templates/apiExtractor.ts
@@ -1,0 +1,4 @@
+export const apiExtractor = {
+  $schema: 'https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json',
+  extends: '@fluentui/scripts/api-extractor/api-extractor.common.json',
+};
