+++ tools/generators/migrate-converged-pkg/templates/babelConfig.ts
@@ -1,0 +1,5 @@
+export const babelConfig = (options: { extraPlugins: Array<string> }) => {
+  return {
+    plugins: [...options.extraPlugins, 'annotate-pure-calls', '@babel/transform-react-pure-annotations'],
+  };
+};
