+++ tools/generators/migrate-converged-pkg/templates/index.ts
@@ -1,0 +1,9 @@
+export * from './apiExtractor';
+export * from './apiExtractorLocal';
+export * from './babelConfig';
+export * from './e2e';
+export * from './jestConfig';
+export * from './jestSetup';
+export * from './npmignore';
+export * from './storybook';
+export * from './tsconfig';
