+++ tools/generators/migrate-converged-pkg/templates/jestSetup.ts
@@ -1,0 +1,5 @@
+import { stripIndents } from '@nrwl/devkit';
+
+export const jestSetup = stripIndents`
+/** Jest test setup file. */
+`;
