+++ tools/generators/migrate-converged-pkg/updaters/index.ts
@@ -1,0 +1,11 @@
+export * from './updateApiExtractorForLocalBuilds';
+export * from './updateBabel';
+export * from './updateBaseTsConfig';
+export * from './updateE2E';
+export * from './updateLocalJestConfig';
+export * from './updateLocalTsConfig';
+export * from './updateNpmIgnore';
+export * from './updateNpmScripts';
+export * from './updateNxWorkspace';
+export * from './updateRootJestConfig';
+export * from './updateStorybook';
