+++ tools/generators/migrate-converged-pkg/updaters/updateRootJestConfig.ts
@@ -1,0 +1,9 @@
+import { Tree } from '@nrwl/devkit';
+import { updateJestConfig } from '../../../utils';
+import { NormalizedSchema } from '../types';
+
+export function updateRootJestConfig(tree: Tree, options: NormalizedSchema) {
+  updateJestConfig(tree, { project: options.name });
+
+  return tree;
+}
