+++ tools/generators/migrate-converged-pkg/utils/uniqueArray.ts
@@ -1,0 +1,3 @@
+export function uniqueArray<T extends unknown>(value: T[]) {
+  return Array.from(new Set(value));
+}
